refactor: split next wave of God-Files
Thin wrappers for references/new actions; decompose companies-grid, reference detail page, and table column header into focused modules.
